little bit of cleanup, organizing
